refactor: make scan method more legible
